+++ src/EntityFramework.Storage/test/IntegrationTests/TokenCleanup/TokenCleanupTests.cs
@@ -30,7 +30,7 @@ public class TokenCleanupTests : IntegrationTest<TokenCleanupTests, PersistedGra
 
     public TokenCleanupTests(DatabaseProviderFixture<PersistedGrantDbContext> fixture) : base(fixture)
     {
-        foreach (var options in TestDatabaseProviders.SelectMany(x => x.Select(y => (DbContextOptions<PersistedGrantDbContext>)y)).ToList())
+        foreach (var options in TestDatabaseProviders)
         {
             using (var context = new PersistedGrantDbContext(options, StoreOptions))
             {