+++ src/EntityFramework/test/IdentityServer8.EntityFramework.Tests/IntegrationTest.cs
@@ -60,7 +60,7 @@ public class IntegrationTest<TClass, TDbContext, TStoreOption> : IClassFixture<D
 
     protected IntegrationTest(DatabaseProviderFixture<TDbContext> fixture)
     {
-        fixture.Options = TestDatabaseProviders.SelectMany(x => x.Select(y => (DbContextOptions<TDbContext>)y)).ToList();
+        fixture.Options = (TestDatabaseProviders as IEnumerable<DbContextOptions<TDbContext>>).ToList();
         fixture.StoreOptions = StoreOptions;
     }
 }