+++ src/IdentityServer8/src/Validation/Default/TokenValidator.cs
@@ -10,6 +10,7 @@
  copies or substantial portions of the Software.
 */
 
+using Microsoft.IdentityModel.JsonWebTokens;
 using Microsoft.IdentityModel.Tokens;
 
 namespace IdentityServer8.Validation;
@@ -237,7 +238,7 @@ internal class TokenValidator : ITokenValidator
 
     private async Task<TokenValidationResult> ValidateJwtAsync(string jwt, IEnumerable<SecurityKeyInfo> validationKeys, bool validateLifetime = true, string audience = null)
     {
-        var handler = new JwtSecurityTokenHandler();
+        var handler = new JsonWebTokenHandler();
         handler.InboundClaimTypeMap.Clear();
 
         var parameters = new TokenValidationParameters
@@ -258,15 +259,17 @@ internal class TokenValidator : ITokenValidator
 
         try
         {
-            var id = handler.ValidateToken(jwt, parameters, out var securityToken);
-            var jwtSecurityToken = securityToken as JwtSecurityToken;
+            var result = await handler.ValidateTokenAsync(jwt, parameters);
+            var securityToken = result.SecurityToken;
+            var id = result.ClaimsIdentity;
+            var jwtSecurityToken = securityToken as JsonWebToken;
 
             // if no audience is specified, we make at least sure that it is an access token
             if (audience.IsMissing())
             {
                 if (_options.AccessTokenJwtType.IsPresent())
                 {
-                    var type = jwtSecurityToken.Header.Typ;
+                    var type = jwtSecurityToken.Typ;
                     if (!string.Equals(type, _options.AccessTokenJwtType))
                     {
                         return new TokenValidationResult
@@ -405,7 +408,7 @@ internal class TokenValidator : ITokenValidator
     {
         try
         {
-            var jwt = new JwtSecurityToken(token);
+            var jwt = new JsonWebToken(token);
             var clientId = jwt.Audiences.FirstOrDefault();
 
             return clientId;