+++ src/IdentityServer8/test/IdentityServer.IntegrationTests/Endpoints/Introspection/IntrospectionTests.cs
@@ -188,7 +188,7 @@ public class IntrospectionTests
             Token = tokenResponse.AccessToken
         });
 
-        var values = introspectionResponse.Json.ToObject<Dictionary<string, JsonElement>>();
+        var values = introspectionResponse.Json?.Deserialize<Dictionary<string, JsonElement>>() ?? [];
 
         values["aud"].ValueKind.Should().Be(JsonValueKind.String);
         values["iss"].ValueKind.Should().Be(JsonValueKind.String);
@@ -227,7 +227,7 @@ public class IntrospectionTests
             Token = tokenResponse.AccessToken
         });
 
-        var values = introspectionResponse.Json.ToObject<Dictionary<string, JsonElement>>();
+        var values = introspectionResponse.Json?.Deserialize<Dictionary<string, JsonElement>>() ?? [];
 
         values["aud"].ValueKind.Should().Be(JsonValueKind.String);
         values["iss"].ValueKind.Should().Be(JsonValueKind.String);
@@ -266,7 +266,7 @@ public class IntrospectionTests
             Token = tokenResponse.AccessToken
         });
 
-        var values = introspectionResponse.Json.ToObject<Dictionary<string, JsonElement>>();
+        var values = introspectionResponse.Json?.Deserialize<Dictionary<string, JsonElement>>() ?? [];
 
 
         values["aud"].GetType().Name.Should().Be("JsonElement");
@@ -317,7 +317,7 @@ public class IntrospectionTests
             Token = tokenResponse.AccessToken
         });
 
-        var values = introspectionResponse.Json.ToObject<Dictionary<string, JsonElement>>();
+        var values = introspectionResponse.Json?.Deserialize<Dictionary<string, JsonElement>>() ?? [];
         values["aud"].ValueKind.Should().Be(JsonValueKind.String);
         values["aud"].ValueKind.Should().Be(JsonValueKind.String);
         values["iss"].ValueKind.Should().Be(JsonValueKind.String);